Develop

Closes S03P13B107-110 and S03P13B107-20

See merge request s03-webmobile1-sub3/s03p13b107!188

+++ frontend/src/views/accounts/DeleteUserView.vue
@@ -1,5 +1,5 @@
 <template>
-  <div>회원탈퇴</div>
+  <div style="font-family: 'Do Hyeon', sans-serif;">회원탈퇴</div>
 </template>
 
 <script>
@@ -8,12 +8,17 @@ import { mapActions } from "vuex";
 
 @Component({
   methods: {
-    ...mapActions(["deleteUser"]),
-  },
+    ...mapActions(["deleteUser"])
+  }
 })
 export default class DeleteUserView extends Vue {
   mounted() {
-    this.deleteUser();
+    const check = confirm("정말 탈퇴하시겠습니까?");
+    if (check == true) {
+      this.deleteUser();
+    } else {
+      this.$router.push("/accounts/update");
+    }
   }
 }
 </script>
